+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Added commenting @robgietema
 * Added tags @robgietema
 * Upgrade to React 16 @sneridagh
+* Upgrade to Webpack 4 @sneridagh
 
 ## 0.4.0 (2017-05-03)